chore: generate int64 instead of int in client types
kube generally avoids int, because it is architecture dependent (which
is not great for an API), and uses int64 instead.

+++ scripts/generate-go-crd-clients/client_test.go
@@ -119,7 +119,7 @@ func TestComputeInstanceGoClient(t *testing.T) {
 
 	computeInstanceName := "computeinstance"
 	autoDelete := true
-	bootDiskSize := 20
+	bootDiskSize := int64(20)
 	bootDiskType := "pd-ssd"
 	machineType := "n1-standard-1"
 	zone := "europe-west1-b"
